move this class ValidatePagesService to the package com.topcoder.productsearch.cleaner.service
