+++ crawler-service/src/main/java/com/topcoder/productsearch/common/ProcessLauncher.java
@@ -5,7 +5,7 @@ import com.topcoder.productsearch.common.entity.WebSite;
 import com.topcoder.productsearch.common.repository.WebSiteRepository;
 import com.topcoder.productsearch.converter.service.ConverterService;
 import com.topcoder.productsearch.crawler.service.CrawlerService;
-import com.topcoder.productsearch.validatePages.service.ValidatePagesService;
+import com.topcoder.productsearch.cleaner.service.ValidatePagesService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
