+++ crawler-service/src/test/java/com/topcoder/productsearch/common/ProcessLauncherTest.java
@@ -6,7 +6,7 @@ import com.topcoder.productsearch.common.entity.WebSite;
 import com.topcoder.productsearch.common.repository.WebSiteRepository;
 import com.topcoder.productsearch.converter.service.ConverterService;
 import com.topcoder.productsearch.crawler.service.CrawlerService;
-import com.topcoder.productsearch.validatePages.service.ValidatePagesService;
+import com.topcoder.productsearch.cleaner.service.ValidatePagesService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
